split test graph library validation fixtures
